+++ cli/src/pcluster/validators/login_nodes_validators.py
@@ -20,10 +20,11 @@ class LoginNodesSshKeyNameDeprecatedValidator(Validator):
     """Validator to warn about deprecated LoginNodes SSH KeyName."""
 
     def _validate(self, key_name: str = None, key_name_explicitly_set: bool = False):
-        """Validate that LoginNodes SSH KeyName is deprecated."""
+        """Return a deprecation warning when LoginNodes SSH KeyName is explicitly set in the cluster config."""
         if key_name and key_name_explicitly_set:
             self._add_failure(
-                "LoginNodes/Pools/Ssh/KeyName is deprecated. Please remove it from cluster configuration. "
+                "LoginNodes/Pools/Ssh/KeyName is deprecated since ParallelCluster version 3.14.0. "
+                "Please remove it from cluster configuration. "
                 "The SSH key for the login nodes will be the one used by the head node.",
                 FailureLevel.WARNING,
             )